refactor naming to merch store
